feat mig models db
