+++ app/Http/Controllers/PostController.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use App\Models\Post;
+use Illuminate\Http\Request;
+
+class PostController extends Controller
+{
+    public function index()
+    {
+        return view('posts', ["title" => "Blog", "posts" => Post::all()]);
+    }
+
+    public function show($slug)
+    {
+        return view('post', [
+            "title" => "Single Post",
+            "post" => Post::find($slug)
+        ]);
+    }
+}
