+++ app/Models/Post.php
@@ -2,31 +2,9 @@
 
 namespace App\Models;
 
-class Post
-{
-    private  static $blog_posts = [
-        [
-            "title" => "Judul Post Pertama",
-            "slug" => "judul-post-pertama",
-            "author" => "Oji",
-            "body" => "Lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quod. Lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quod."
-        ],
-        [
-            "title" => "Judul Post Kedua",
-            "slug" => "judul-post-kedua",
-            "author" => "Rizky",
-            "body" => "Lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quod. Lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quod."
-        ],
-        [
-            "title" => "Judul Post Ketiga",
-            "slug" => "judul-post-ketiga",
-            "author" => "Rahmat",
-            "body" => "Lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quod. Lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quod."
-        ]
-    ];
+use Illuminate\Database\Eloquent\Model;
 
-    public static function all()
-    {
-        return self::$blog_posts;
-    }
+class Post extends Model
+{
+    //
 }
