+++ routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\PostController;
 use App\Models\Post;
 use Illuminate\Support\Facades\Route;
 
@@ -22,21 +23,22 @@ Route::get('/about', function () {
 
 
 
-Route::get('/blog', function () {
-    return view('posts', ["title" => "Blog", "posts" => Post::all()]);
-});
+Route::get('/blog', [PostController::class, 'index']);
 
-Route::get('posts/{slug}', function ($slug) {
-    $blog_post = Post::all();
-    $new_post = [];
-    foreach ($blog_post as $post) {
-        if ($post["slug"] === $slug) {
-            $new_post = $post;
-        }
-    }
-    //dd($new_post);   
-    return view('post', [
-        "title" => "Single Post",
-        "post" => $new_post
-    ]);
-});
+// Route::get('posts/{slug}', function ($slug) {
+//     // $blog_post = Post::all();
+//     // $new_post = [];
+//     // foreach ($blog_post as $post) {
+//     //     if ($post["slug"] === $slug) {
+//     //         $new_post = $post;
+//     //     }
+//     // }
+//     //dd($new_post);   
+//     return view('post', [
+//         "title" => "Single Post",
+//         // "post" => $new_post,
+//         "post" => Post::find($slug)
+//     ]);
+// });
+
+Route::get('posts/{slug}', [PostController::class, 'show']);
